Split large dispatcher and orchestrator HTTP tests
